Started removing functions from check_known_puslars.py and put them into obs_tools and pulsar_obs_helper
