+++ scripts/bestgridpos.py
@@ -7,7 +7,7 @@ from math import cos,sin
 import glob
 import sys
 
-from check_known_pulsars import calc_ta_fwhm
+from dpp.obs_tools import calc_ta_fwhm
 from mwa_metadb_utils import get_common_obs_metadata, get_obs_array_phase
 from submit_to_database import sex2deg
 from find_pulsar_in_obs import deg2sex, format_ra_dec
